final-project/requirements.txt
flask
python-dotenv
selenium
bs4
pytest
